Added dominant colour and masking methods
Written method to find the dominant colour (based on an algorithm
proposed in Ekin et al (2003) "Automatic Soccer Video Analysis and
Summarization") and method to find the standard deviation in a colour
channel based on this dominant colour.
Written method which aims to use the dominant colour and standard
deviation to create and apply a mask to the image.
These will eventually be used to mask the football pitch on the dominant
colour (the grass green).