Updated dominant colour mean and deviation methods
Added ability to find multiple dominant colours in an image channel.
This is necessary for images with high contrast in the football field,
where a single dominant colour is insufficient.
Added ability to define a range to find the dominant colour within. This
will allow the algorithm to focus on the green region of the hue
channel.
Updated docstrings.
Converted dimensions in COLOUR_SPACES to ndarrays.
Corrected multi-line strings.