Modified create_mask to accept multiple colours
Rewritten create_mask() method to use the new dominant colour and
dominant colour deviation algorithms. Function now takes in an arbitrary
number of colours (and their standard deviations) and creates a mask of
the image where any of those colours are present.
This is essentially a mixture of Gaussians model, and will be used for
images with large shadows.